diassembler x86: Fix eip-relative addressing, XOP support and MOVBE/CRC32
When an address size override is used together with RIP-relative
addressing the result should be that the address is truncated, not
that it's decoded the same as it would be in 32-bit mode.

AMD's XOP-instructions are like the VEX instructions, but on a
different opcode byte. The XOP opcode tables tables are also not
anywhere inside the regular opcode table. Almost all XOP instructions
are included. The old operand syntax that was used for AVX will be
migrated to AMD's operand syntax.

The MOVBE/CRC32 opcode is weird. It looks like an SSE encoded
instruction, except that #x66 is an operand size override and does not
select another instruction. This requires special handling.