Disassemblers can now compute PC-relative addressing
The disassemblers were previous not provided with the program counter,
so all PC-relative addressing (in e.g. branches) was returned as
symbolic expressions. Now the current program counter can be provided
to the disassembler for improved output.

(This was mainly motived by ARM A64, which has a page-aligned
addressing mode that just looks weird when written symbolically).